% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runPower.R
\name{runPower}
\alias{runPower}
\title{Power Analysis for Causal Mediation Analysis}
\usage{
runPower(
  seed_user = 12,
  n = 100,
  num_x = 2,
  treat.prop = 0.5,
  treat.randomized = FALSE,
  R2.ax = NULL,
  R2.mx = NULL,
  R2.yx = NULL,
  std.m_on_a = NULL,
  std.y_on_a = NULL,
  std.y_on_m = NULL,
  std.y_on_am_2way = NULL,
  std.y_on_m_2way = NULL,
  std.y_on_am_3way = NULL,
  a_on_x = sqrt(0.13),
  m_on_a = rep(0.14, 2),
  m_on_x = rep(sqrt(0.13), 2),
  em_corr = 0,
  M_binary = rep(FALSE, 2),
  y_on_a = 0.1,
  y_on_m = rep(0.39, 2),
  y_on_am_2way = rep(0.1, 2),
  y_on_m_2way = rep(0.1, 2 * (2 - 1)/2),
  y_on_am_3way = rep(0, 2 * (2 - 1)/2),
  y_on_x = sqrt(0.02),
  Y_binary = FALSE,
  nboot = NULL,
  n.draws = 1000,
  nreps.binYM = 1000,
  sig.level = 0.05,
  nsims = 1000,
  mc.cores = 1
)
}
\arguments{
\item{n}{Sample size.}
}
\value{
A \code{data.frame} containing the Type I error rates and statistical power for testing interventional indirect effects via each mediator. Specifically, the function outputs (1) the familywise ("FW") Type I error rate and power of testing all four indirect effects for a mediator (i.e., the probability of finding significance for at least one of the indirect effects for a mediator, when that indirect effect were actually null and non-null, respectively), and (2) the per-test ("PT") Type I error rate and power of testing only one indirect effect for a mediator.
}
\description{
Implement a simulation-based method for statistical power calculation in causal mediation analysis with two correlated mediators.
}
